step 12: reconstruct the Derge layout using 3-1_reinsert_page_nums.py
Here, check if all "a" were inserted correctly, looking at the beginning page in Derge and last page.
The easiest is to go at the end of a Derge line and scroll down until the end of the file, so as to see lines with a double size. It shows that an "a" was missing there.
Also look at the output in the console, as it will give indications of what went wrong for which file.

In order correct the "a"s:
      - modify the files in 2-0-with_a
      - rerun 2-1_reinsert_a.py (uses the Linux diff command, so must be executed in a Linux OS)
      - rerun 3-1_reinsert_page_nums.py
      - check if everything is ok.